Added success session flash messages. User can now edit and delete posts he owns, and can edit his profile.

+++ views/posts/show.php
@@ -6,7 +6,6 @@
 $this->title = 'Posts'
 ?>
 <div class="container">
-    <h2>ID: <?= $post->id ?></h2>
     <h2>Title: <?= $post->title ?></h2>
     <h2>Body: <?= $post->body ?></h2>
     <h2>Author: <?= $users->findOne(['id' => $post->user_id])->username ?></h2>
@@ -17,8 +16,12 @@ $this->title = 'Posts'
                 <a href="/posts/approve?id=<?= $post->id ?>&approved=false"><button type="submit" class="btn btn-secondary">Unapprove</button></a>
             <?php else: ?>
                 <a href="/posts/approve?id=<?= $post->id ?>&approved=true"><button type="submit" class="btn btn-success">Approve</button></a>
-            <?php endif; ?></h2>
-    <a href="/posts/edit?id=<?= $post->id ?>"><button class="btn btn-primary mr-3">Edit</button></a>
-    <a href="/posts/delete?id=<?= $post->id ?>"><button class="btn btn-danger">Delete</button></a>
-    <?php endif; ?>
+            <?php endif;
+            endif;
+            ?></h2>
+        <?php if ($users->isAdmin() || $post->user_id === $users->getId()): ?>
+                <a href="/posts/edit?id=<?= $post->id ?>"><button class="btn btn-primary mr-3">Edit</button></a>
+                <a href="/posts/delete?id=<?= $post->id ?>"><button class="btn btn-danger">Delete</button></a>
+            <?php
+        endif; ?>
 </div>
